+++ docsig/messages.py
@@ -1,0 +1,7 @@
+"""
+docsig.messages
+===============
+"""
+#: Wxxx: Warning
+#: W1xx: Docstring
+W101 = "{module}::{func} is missing a docstring"